[#29] Move raw data category serializer
