+++ backend/api/v1/v1_categories/views.py
@@ -17,11 +17,12 @@ from api.v1.v1_categories.functions import get_category_results
 from api.v1.v1_categories.models import DataCategory
 from api.v1.v1_data.models import FormData, Answers
 from api.v1.v1_forms.models import Forms
-from api.v1.v1_data.serializers import (
-    ListFormDataRequestSerializer,
+from api.v1.v1_categories.serializers import (
     ListRawDataSerializer,
     ListRawDataAnswerSerializer,
 )
+
+from api.v1.v1_data.serializers import ListFormDataRequestSerializer
 
 
 @extend_schema(
